Revert "Move mutators to config methods"
This reverts commit 6f489da5e4166d3b251b85bddb9746a07d6b546b.
