+++ internal/otelcollector/config/log/agent/config_builder.go
@@ -3,6 +3,7 @@ package agent
 import (
 	"context"
 	"fmt"
+	"maps"
 
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -49,7 +50,7 @@ func (b *Builder) Build(ctx context.Context, logPipelines []telemetryv1alpha1.Lo
 			queueSize,
 			otlpexporter.SignalTypeLog,
 		)
-		if err := cfg.addLogPipelineComponents(ctx, otlpExporterBuilder, &pipeline, envVars); err != nil {
+		if err := declareComponentsForLogPipeline(ctx, otlpExporterBuilder, &pipeline, cfg, envVars); err != nil {
 			return nil, nil, err
 		}
 
@@ -57,7 +58,35 @@ func (b *Builder) Build(ctx context.Context, logPipelines []telemetryv1alpha1.Lo
 		cfg.Service.Pipelines[pipelineID] = pipelineConfig(fmt.Sprintf("filelog/%s", pipeline.Name), otlpexporter.ExporterID(pipeline.Spec.Output.OTLP.Protocol, pipeline.Name))
 	}
 
+	// Return the assembled config and any environment variables needed for exporters
 	return cfg, envVars, nil
+}
+
+// declareComponentsForLogPipeline enriches a Config (exporters, processors, etc.) with components for a given telemetryv1alpha1.LogPipeline.
+func declareComponentsForLogPipeline(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.LogPipeline, cfg *Config, envVars otlpexporter.EnvVars) error {
+	declareFileLogReceiver(pipeline, cfg)
+	return declareOTLPExporter(ctx, otlpExporterBuilder, pipeline, cfg, envVars)
+}
+
+func declareFileLogReceiver(pipeline *telemetryv1alpha1.LogPipeline, cfg *Config) {
+	receiver := fileLogReceiverConfig(*pipeline)
+
+	otlpReceiverID := fmt.Sprintf("filelog/%s", pipeline.Name)
+	cfg.Receivers[otlpReceiverID] = Receiver{FileLog: receiver}
+}
+
+func declareOTLPExporter(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.LogPipeline, cfg *Config, envVars otlpexporter.EnvVars) error {
+	otlpExporterConfig, otlpExporterEnvVars, err := otlpExporterBuilder.MakeConfig(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to make otlp exporter config: %w", err)
+	}
+
+	maps.Copy(envVars, otlpExporterEnvVars)
+
+	otlpExporterID := otlpexporter.ExporterID(pipeline.Spec.Output.OTLP.Protocol, pipeline.Name)
+	cfg.Exporters[otlpExporterID] = Exporter{OTLP: otlpExporterConfig}
+
+	return nil
 }
 
 // Each pipeline will have one receiver and one exporter
