+++ internal/otelcollector/config/metric/gateway/config.go
@@ -1,22 +1,14 @@
 package gateway
 
 import (
-	"context"
-	"fmt"
-	"maps"
-
-	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/metric"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/otlpexporter"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/ottlexpr"
-	metricpipelineutils "github.com/kyma-project/telemetry-manager/internal/utils/metricpipeline"
 )
 
 // newConfig constructs a global, pipeline-independent Base config for the metric gateway collector.
 // It sets up default service and extension components, and returns a Config with initialized fields.
 func newConfig(opts BuildOptions) *Config {
-	return &Config{
+	var cfg = Config{
 		Base: config.DefaultBaseConfig(make(config.Pipelines),
 			config.WithK8sLeaderElector("serviceAccount", "telemetry-metric-gateway-kymastats", opts.GatewayNamespace),
 		),
@@ -25,177 +17,8 @@ func newConfig(opts BuildOptions) *Config {
 		Exporters:  make(Exporters),
 		Connectors: make(Connectors),
 	}
-}
 
-// addMetricPipelineComponents enriches a Config (exporters, processors, etc.) with components for a given telemetryv1alpha1.MetricPipeline.
-func (cfg *Config) addMetricPipelineComponents(
-	ctx context.Context,
-	otlpExporterBuilder *otlpexporter.ConfigBuilder,
-	pipeline *telemetryv1alpha1.MetricPipeline,
-	envVars otlpexporter.EnvVars,
-) error {
-	cfg.addDiagnosticMetricsDropFilters(pipeline)
-	cfg.addInputSourceFilters(pipeline)
-	cfg.addRuntimeResourcesFilters(pipeline)
-	cfg.addNamespaceFilters(pipeline)
-	cfg.addConnectors(pipeline.Name)
-
-	return cfg.addOTLPExporter(ctx, otlpExporterBuilder, pipeline, envVars)
-}
-
-func (cfg *Config) addDiagnosticMetricsDropFilters(pipeline *telemetryv1alpha1.MetricPipeline) {
-	input := pipeline.Spec.Input
-
-	if metricpipelineutils.IsPrometheusInputEnabled(input) && !metricpipelineutils.IsPrometheusDiagnosticInputEnabled(input) {
-		cfg.Processors.DropDiagnosticMetricsIfInputSourcePrometheus = dropDiagnosticMetricsFilterConfig(inputSourceEquals(metric.InputSourcePrometheus))
-	}
-
-	if metricpipelineutils.IsIstioInputEnabled(input) && !metricpipelineutils.IsIstioDiagnosticInputEnabled(input) {
-		cfg.Processors.DropDiagnosticMetricsIfInputSourceIstio = dropDiagnosticMetricsFilterConfig(inputSourceEquals(metric.InputSourceIstio))
-	}
-}
-
-func (cfg *Config) addInputSourceFilters(pipeline *telemetryv1alpha1.MetricPipeline) {
-	input := pipeline.Spec.Input
-
-	if !metricpipelineutils.IsRuntimeInputEnabled(input) {
-		cfg.Processors.DropIfInputSourceRuntime = dropIfInputSourceRuntimeProcessorConfig()
-	}
-
-	if !metricpipelineutils.IsPrometheusInputEnabled(input) {
-		cfg.Processors.DropIfInputSourcePrometheus = dropIfInputSourcePrometheusProcessorConfig()
-	}
-
-	if !metricpipelineutils.IsIstioInputEnabled(input) {
-		cfg.Processors.DropIfInputSourceIstio = dropIfInputSourceIstioProcessorConfig()
-	}
-
-	if !metricpipelineutils.IsOTLPInputEnabled(input) {
-		cfg.Processors.DropIfInputSourceOTLP = dropIfInputSourceOTLPProcessorConfig()
-	}
-
-	if !metricpipelineutils.IsIstioInputEnabled(input) || !metricpipelineutils.IsEnvoyMetricsEnabled(input) {
-		cfg.Processors.DropIfEnvoyMetricsDisabled = dropIfEnvoyMetricsDisabledProcessorConfig()
-	}
-}
-
-func (cfg *Config) addRuntimeResourcesFilters(pipeline *telemetryv1alpha1.MetricPipeline) {
-	input := pipeline.Spec.Input
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimePodInputEnabled(input) {
-		cfg.Processors.DropRuntimePodMetrics = dropRuntimePodMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeContainerInputEnabled(input) {
-		cfg.Processors.DropRuntimeContainerMetrics = dropRuntimeContainerMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeNodeInputEnabled(input) {
-		cfg.Processors.DropRuntimeNodeMetrics = dropRuntimeNodeMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeVolumeInputEnabled(input) {
-		cfg.Processors.DropRuntimeVolumeMetrics = dropRuntimeVolumeMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeDeploymentInputEnabled(input) {
-		cfg.Processors.DropRuntimeDeploymentMetrics = dropRuntimeDeploymentMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeStatefulSetInputEnabled(input) {
-		cfg.Processors.DropRuntimeStatefulSetMetrics = dropRuntimeStatefulSetMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeDaemonSetInputEnabled(input) {
-		cfg.Processors.DropRuntimeDaemonSetMetrics = dropRuntimeDaemonSetMetricsProcessorConfig()
-	}
-
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && !metricpipelineutils.IsRuntimeJobInputEnabled(input) {
-		cfg.Processors.DropRuntimeJobMetrics = dropRuntimeJobMetricsProcessorConfig()
-	}
-}
-
-func (cfg *Config) addNamespaceFilters(pipeline *telemetryv1alpha1.MetricPipeline) {
-	if cfg.Processors.NamespaceFilters == nil {
-		cfg.Processors.NamespaceFilters = make(NamespaceFilters)
-	}
-
-	input := pipeline.Spec.Input
-	if metricpipelineutils.IsRuntimeInputEnabled(input) && shouldFilterByNamespace(input.Runtime.Namespaces) {
-		processorID := formatNamespaceFilterID(pipeline.Name, metric.InputSourceRuntime)
-		cfg.Processors.NamespaceFilters[processorID] = filterByNamespaceProcessorConfig(pipeline.Spec.Input.Runtime.Namespaces, inputSourceEquals(metric.InputSourceRuntime))
-	}
-
-	if metricpipelineutils.IsPrometheusInputEnabled(input) && shouldFilterByNamespace(input.Prometheus.Namespaces) {
-		processorID := formatNamespaceFilterID(pipeline.Name, metric.InputSourcePrometheus)
-		cfg.Processors.NamespaceFilters[processorID] = filterByNamespaceProcessorConfig(pipeline.Spec.Input.Prometheus.Namespaces, ottlexpr.ResourceAttributeEquals(metric.KymaInputNameAttribute, metric.KymaInputPrometheus))
-	}
-
-	if metricpipelineutils.IsIstioInputEnabled(input) && shouldFilterByNamespace(input.Istio.Namespaces) {
-		processorID := formatNamespaceFilterID(pipeline.Name, metric.InputSourceIstio)
-		cfg.Processors.NamespaceFilters[processorID] = filterByNamespaceProcessorConfig(pipeline.Spec.Input.Istio.Namespaces, inputSourceEquals(metric.InputSourceIstio))
-	}
-
-	if metricpipelineutils.IsOTLPInputEnabled(input) && input.OTLP != nil && shouldFilterByNamespace(input.OTLP.Namespaces) {
-		processorID := formatNamespaceFilterID(pipeline.Name, metric.InputSourceOTLP)
-		cfg.Processors.NamespaceFilters[processorID] = filterByNamespaceProcessorConfig(pipeline.Spec.Input.OTLP.Namespaces, otlpInputSource())
-	}
-}
-
-func (cfg *Config) addConnectors(pipelineName string) {
-	forwardConnectorID := formatForwardConnectorID(pipelineName)
-	cfg.Connectors[forwardConnectorID] = struct{}{}
-
-	routingConnectorID := formatRoutingConnectorID(pipelineName)
-	cfg.Connectors[routingConnectorID] = routingConnectorConfig(pipelineName)
-}
-
-func (cfg *Config) addOTLPExporter(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.MetricPipeline, envVars otlpexporter.EnvVars) error {
-	otlpExporterConfig, otlpExporterEnvVars, err := otlpExporterBuilder.MakeConfig(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to make otlp exporter config: %w", err)
-	}
-
-	maps.Copy(envVars, otlpExporterEnvVars)
-
-	exporterID := otlpexporter.ExporterID(pipeline.Spec.Output.OTLP.Protocol, pipeline.Name)
-	cfg.Exporters[exporterID] = Exporter{OTLP: otlpExporterConfig}
-
-	return nil
-}
-
-func (cfg *Config) addInputPipeline(pipelineName string) {
-	cfg.Service.Pipelines[formatInputPipelineID(pipelineName)] = config.Pipeline{
-		Receivers:  []string{"otlp", "kymastats"},
-		Processors: []string{"memory_limiter"},
-		Exporters:  []string{formatRoutingConnectorID(pipelineName)},
-	}
-}
-
-func (cfg *Config) addEnrichmentPipeline(pipelineName string) {
-	cfg.Service.Pipelines[formatAttributesEnrichmentPipelineID(pipelineName)] = config.Pipeline{
-		Receivers:  []string{formatRoutingConnectorID(pipelineName)},
-		Processors: []string{"k8sattributes", "service_enrichment"},
-		Exporters:  []string{formatForwardConnectorID(pipelineName)},
-	}
-}
-
-func (cfg *Config) addOutputPipeline(pipeline *telemetryv1alpha1.MetricPipeline) {
-	input := pipeline.Spec.Input
-
-	processors := []string{}
-	processors = append(processors, "transform/set-instrumentation-scope-kyma")
-	processors = append(processors, inputSourceFiltersIDs(input)...)
-	processors = append(processors, namespaceFiltersIDs(input, pipeline)...)
-	processors = append(processors, runtimeResourcesFiltersIDs(input)...)
-	processors = append(processors, diagnosticMetricFiltersIDs(input)...)
-	processors = append(processors, "resource/insert-cluster-attributes", "resource/delete-skip-enrichment-attribute", "resource/drop-kyma-attributes", "batch")
-
-	cfg.Service.Pipelines[formatOutputPipelineID(pipeline.Name)] = config.Pipeline{
-		Receivers:  []string{formatRoutingConnectorID(pipeline.Name), formatForwardConnectorID(pipeline.Name)},
-		Processors: processors,
-		Exporters:  []string{formatOTLPExporterID(pipeline)},
-	}
+	return &cfg
 }
 
 type Config struct {
