+++ internal/otelcollector/config/trace/gateway/config.go
@@ -1,13 +1,7 @@
 package gateway
 
 import (
-	"context"
-	"fmt"
-	"maps"
-
-	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/otlpexporter"
 )
 
 // newConfig constructs a global, pipeline-independent Base config for the metric agent collector.
@@ -22,25 +16,6 @@ func newConfig(opts BuildOptions) *Config {
 		Processors: processorsConfig(opts),
 		Exporters:  make(Exporters),
 	}
-}
-
-// declareComponentsForTracePipeline enriches a Config (exporters, processors, etc.) with components for a given telemetryv1alpha1.TracePipeline.
-func (cfg *Config) addTracePipelineComponents(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.TracePipeline, envVars otlpexporter.EnvVars) error {
-	return cfg.addOTLPExporter(ctx, otlpExporterBuilder, pipeline, envVars)
-}
-
-func (cfg *Config) addOTLPExporter(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.TracePipeline, envVars otlpexporter.EnvVars) error {
-	otlpExporterConfig, otlpExporterEnvVars, err := otlpExporterBuilder.MakeConfig(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to make otlp exporter config: %w", err)
-	}
-
-	maps.Copy(envVars, otlpExporterEnvVars)
-
-	otlpExporterID := otlpexporter.ExporterID(pipeline.Spec.Output.OTLP.Protocol, pipeline.Name)
-	cfg.Exporters[otlpExporterID] = Exporter{OTLP: otlpExporterConfig}
-
-	return nil
 }
 
 type Config struct {
