--- internal/otelcollector/config/trace/gateway/receivers.go
@@ -1,23 +1,0 @@
-package gateway
-
-import (
-	"fmt"
-
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/ports"
-)
-
-func receiversConfig() Receivers {
-	return Receivers{
-		OTLP: config.OTLPReceiver{
-			Protocols: config.ReceiverProtocols{
-				HTTP: config.Endpoint{
-					Endpoint: fmt.Sprintf("${%s}:%d", config.EnvVarCurrentPodIP, ports.OTLPHTTP),
-				},
-				GRPC: config.Endpoint{
-					Endpoint: fmt.Sprintf("${%s}:%d", config.EnvVarCurrentPodIP, ports.OTLPGRPC),
-				},
-			},
-		},
-	}
-}
